pyproject.toml
[tool.poetry]
name = "structconf"
version = "0.2.1"
description = "struct config using yaml"
authors = ["Wenqi Zhao <1027572886a@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/Green-Wood/structconf"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "^3.7"
pydantic = "^1.8.1"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
